requirements.txt
openai>=1.12.0
requests>=2.31.0
python-dotenv>=1.0.0
lark-oapi>=1.2.0
schedule>=1.2.0
icalendar>=5.0.0
google-api-python-client>=2.0.0
google-auth>=2.0.0

# creative bot
pyyaml>=6.0

# sentiment bot
jieba>=0.42.1
apscheduler>=3.10.0
PyGithub>=2.1.0

# newsbot
feedparser>=6.0.0
beautifulsoup4>=4.12.0
lxml>=5.0.0

# conductor
volcengine-python-sdk[ark]>=5.0.0   # Seedance 视频 + Seedream 图片生成
# playwright>=1.40.0                # 取消注释启用自动发布（使用系统 Chrome，无需 playwright install）
